+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 2.4.0
 
 * Added Disposable and its related extensions.
+* Fix onCancel not triggering immediately when cancelled.
 
 ## 2.3.0